Add host cert volume and replace X display with more compatible ":0.0" statement
I have noticed that the host certs on all containers are now the same
to avoid any risk from this I have added a volume for the generated
host keys which should be used to add user specific host keys
which are taken over with each new version.

In addition I have noticed that the "localhost:0.0" X display var
is clashing with some Windows security guidelines so I have replaced
it with ":0.0" which seems to be more ok from a Windows point of view.